Move vote simulator to separate module
